genetlink, genltest: refactor and use genltest
